backend/requirements.txt
# backend/requirements.txt
Flask
Flask-Cors
python-dotenv
supabase
paystackease
hypercorn
gunicorn